+++ src/index.tsx
@@ -45,7 +45,7 @@ import IdFind from './pages/AuthFind/IdFind/IdFind';
 import PasswordFind from './pages/AuthFind/PasswordFind/PasswordFind';
 import KaKaoRegister from './pages/OauthRegister/KaKaoRegister/KaKaoRegister';
 import NaverRegister from './pages/OauthRegister/NaverRegister/NaverRegister';
-import Social from './pages/SocialLogin/Social';
+import NaverSocial from './pages/SocialLogin/NaverSocial';
 
 const router = createBrowserRouter([
 	{
@@ -59,7 +59,7 @@ const router = createBrowserRouter([
 			},
 			{
 				path: 'socialLogin',
-				element: <Social />,
+				element: <NaverSocial />,
 			},
 			{
 				path: 'idFind',
@@ -78,7 +78,7 @@ const router = createBrowserRouter([
 				element: <KaKaoRegister />,
 			},
 			{
-				path: 'naverRegister',
+				path: 'oauthAddInfo',
 				element: <NaverRegister />,
 			},
 			{
